+++ static/gsApp/components/features/insightsDateRangeQueryLimitFooter.spec.tsx
@@ -2,10 +2,10 @@ import {OrganizationFixture} from 'sentry-fixture/organization';
 
 import {BillingConfigFixture} from 'getsentry-test/fixtures/billingConfig';
 import {SubscriptionFixture} from 'getsentry-test/fixtures/subscription';
+import {PlanTier} from 'getsentry-test/planTier';
 import {render, screen} from 'sentry-test/reactTestingLibrary';
 
 import {InsightsDateRangeQueryLimitFooter} from 'getsentry/components/features/insightsDateRangeQueryLimitFooter';
-import {PlanTier} from 'getsentry/types';
 
 describe('InsightsUpsellPage', () => {
   const organization = OrganizationFixture();
