+++ static/gsApp/components/features/planFeature.spec.tsx
@@ -3,11 +3,11 @@ import {OrganizationFixture} from 'sentry-fixture/organization';
 import {BillingConfigFixture} from 'getsentry-test/fixtures/billingConfig';
 import {PlanDetailsLookupFixture} from 'getsentry-test/fixtures/planDetailsLookup';
 import {SubscriptionFixture} from 'getsentry-test/fixtures/subscription';
+import {PlanTier} from 'getsentry-test/planTier';
 import {render, waitFor} from 'sentry-test/reactTestingLibrary';
 
 import PlanFeature from 'getsentry/components/features/planFeature';
 import {SubscriptionStore} from 'getsentry/stores/subscriptionStore';
-import {PlanTier} from 'getsentry/types';
 
 describe('PlanFeature', () => {
   const organization = OrganizationFixture();
