+++ static/gsApp/components/productSelectionAvailability.spec.tsx
@@ -2,6 +2,7 @@ import {OrganizationFixture} from 'sentry-fixture/organization';
 
 import {BillingConfigFixture} from 'getsentry-test/fixtures/billingConfig';
 import {SubscriptionFixture} from 'getsentry-test/fixtures/subscription';
+import {PlanTier} from 'getsentry-test/planTier';
 import {
   act,
   render,
@@ -20,7 +21,6 @@ import {ProductSelectionAvailability} from 'getsentry/components/productSelectio
 import type {Reservations} from 'getsentry/components/upgradeNowModal/types';
 import {usePreviewData} from 'getsentry/components/upgradeNowModal/usePreviewData';
 import {SubscriptionStore} from 'getsentry/stores/subscriptionStore';
-import {PlanTier} from 'getsentry/types';
 
 jest.mock('getsentry/components/upgradeNowModal/usePreviewData');
