+++ static/gsApp/components/upgradeNowModal/useUpgradeNowParams.spec.tsx
@@ -3,9 +3,8 @@ import {OrganizationFixture} from 'sentry-fixture/organization';
 import {BillingConfigFixture} from 'getsentry-test/fixtures/billingConfig';
 import {PlanDetailsLookupFixture} from 'getsentry-test/fixtures/planDetailsLookup';
 import {SubscriptionFixture} from 'getsentry-test/fixtures/subscription';
+import {PlanTier} from 'getsentry-test/planTier';
 import {renderHookWithProviders, waitFor} from 'sentry-test/reactTestingLibrary';
-
-import {PlanTier} from 'getsentry/types';
 
 import {useUpgradeNowParams} from './useUpgradeNowParams';
