+++ static/gsApp/views/subscriptionPage/usageOverview/components/charts.spec.tsx
@@ -4,13 +4,13 @@ import {BillingStatFixture} from 'getsentry-test/fixtures/billingStat';
 import {CustomerUsageFixture} from 'getsentry-test/fixtures/customerUsage';
 import {SubscriptionFixture} from 'getsentry-test/fixtures/subscription';
 import {UsageTotalFixture} from 'getsentry-test/fixtures/usageTotal';
+import {PlanTier} from 'getsentry-test/planTier';
 import {act, render, screen} from 'sentry-test/reactTestingLibrary';
 
 import {DataCategory} from 'sentry/types/core';
 import {OrganizationContext} from 'sentry/utils/organizationContext';
 
 import {SubscriptionStore} from 'getsentry/stores/subscriptionStore';
-import {PlanTier} from 'getsentry/types';
 import {UsageCharts} from 'getsentry/views/subscriptionPage/usageOverview/components/charts';
 import type {BreakdownPanelProps} from 'getsentry/views/subscriptionPage/usageOverview/types';
