+++ tests/js/getsentry-test/fixtures/planDetailsLookup.ts
@@ -5,8 +5,7 @@ import {AM2_PLANS} from 'getsentry-test/fixtures/am2Plans';
 import {AM3_PLANS} from 'getsentry-test/fixtures/am3Plans';
 import {MM1_PLANS} from 'getsentry-test/fixtures/mm1Plans';
 import {MM2_PLANS} from 'getsentry-test/fixtures/mm2Plans';
-
-import {PlanTier} from 'getsentry/types';
+import {PlanTier} from 'getsentry-test/planTier';
 
 export type PlanIds =
   | keyof typeof AM1_PLANS
